5316: Removed duplicate file
